@@ -1,5 +1,5 @@
 <?php
 
-function autowire(string $class = null): \WPDesk\ShopMagic\DI\AutowireDefinitionHelper {
+function autowire(?string $class = null): \WPDesk\ShopMagic\DI\AutowireDefinitionHelper {
 	return new \WPDesk\ShopMagic\DI\AutowireDefinitionHelper($class);
 }
