@@ -35,7 +35,7 @@ class ParameterBag implements \IteratorAggregate, \Countable {
 	 *
 	 * @param string|null $key The name of the parameter to return or null to get them all
 	 */
-	public function all( string $key = null ): array {
+	public function all( ?string $key = null ): array {
 		if ( null === $key ) {
 			return $this->parameters;
 		}
