@@ -27,7 +27,7 @@ final class FailingLanguageValidator extends WorkflowValidator {
 		$this->logger  = $logger;
 	}
 
-	public function valid( DataLayer $resources = null ): bool {
+	public function valid( ?DataLayer $resources = null ): bool {
 		$resources = $resources ?? $this->resources;
 		if ( $resources === null ) {
 			return false;
