@@ -37,14 +37,14 @@ final class Util
                 $curlErrno = curl_errno($ch);
                 if (\false === in_array($curlErrno, self::$retriableErrorCodes, \true) || !$retries) {
                     $curlError = curl_error($ch);
-                    if ($closeAfterDone) {
+                    if (\PHP_VERSION_ID < 80000 && $closeAfterDone) {
                         curl_close($ch);
                     }
                     throw new \RuntimeException(sprintf('Curl error (code %d): %s', $curlErrno, $curlError));
                 }
                 continue;
             }
-            if ($closeAfterDone) {
+            if (\PHP_VERSION_ID < 80000 && $closeAfterDone) {
                 curl_close($ch);
             }
             return $curlResponse;
