@@ -22,7 +22,11 @@ class Shop
     /**
      * @var array<string, string>
      */
-    private $shops_usage_tracking_pages = ['wpdesk.pl' => 'https://www.wpdesk.pl/dane-uzytkowania/', 'wpdesk.net' => 'https://www.wpdesk.net/usage-tracking/', 'octolize.com' => 'https://octolize.com/usage-tracking/', 'shopmagic.app' => 'https://www.shopmagic.app/usage-tracking/', 'flexibleinvoices.com' => 'https://www.flexibleinvoices.com/usage-tracking/'];
+    private $shops_usage_tracking_pages = ['wpdesk.pl' => 'https://www.wpdesk.pl/sk/', 'wpdesk.net' => 'https://www.wpdesk.net/sk/', 'octolize.com' => 'https://octolize.com/usage-tracking/', 'shopmagic.app' => 'https://www.shopmagic.app/sk/', 'flexibleinvoices.com' => 'https://www.flexibleinvoices.com/sk/', 'flexiblecoupons.net' => 'https://www.flexiblecoupons.net/sk/'];
+    /**
+     * @var array<string, string>
+     */
+    private $shop_short_slug = ['wpdesk.pl' => 'pl', 'wpdesk.net' => 'net', 'shopmagic.app' => 'sm', 'flexibleinvoices.com' => 'fi', 'flexiblecoupons.net' => 'fc'];
     /**
      * @var array<string, string>
      */
@@ -32,11 +36,17 @@ class Shop
      */
     private $shop;
     /**
-     * @param string $shop_url
+     * @var string
      */
-    public function __construct($shop_url)
+    private $plugin_slug;
+    /**
+     * @param string $shop_url
+     * @param string $plugin_slug
+     */
+    public function __construct($shop_url, $plugin_slug)
     {
         $this->shop = $this->prepare_shop_from_shop_url($shop_url);
+        $this->plugin_slug = $plugin_slug;
     }
     /**
      * @return string
@@ -44,6 +54,10 @@ class Shop
     public function get_usage_tracking_page()
     {
         $usage_tracking_page = isset($this->shops_usage_tracking_pages[$this->shop]) ? $this->shops_usage_tracking_pages[$this->shop] : $this->shops_usage_tracking_pages[$this->default_shop];
+        if ($this->shop !== 'octolize.com') {
+            $shop_shor_slug = isset($this->shop_short_slug[$this->shop]) ? $this->shop_short_slug[$this->default_shop] : 'pl';
+            $usage_tracking_page .= $this->plugin_slug . '-usage-tracking-' . $shop_shor_slug;
+        }
         return apply_filters('wpdesk/tracker/usage_tracking_page', $usage_tracking_page, $this->shop);
     }
     /**
